+++ src/mock/task.js
@@ -1,19 +1,11 @@
+import {COLORS} from "../const.js";
+import {getRandomInteger} from "../utils.js";
+
 const DESCRIPTION = [
   `Изучить теорию`,
   `Сделать домашку`,
   `Пройти интенсив на соточку`
 ];
-
-const COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
-
-// Функция из интернета по генерации случайного числа из диапазона
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
-const getRandomInteger = (min = 0, max = 1) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
 
 const generateDescription = () => {
   const randomIndex = getRandomInteger(0, DESCRIPTION.length - 1);
