+++ src/utils.js
@@ -1,0 +1,35 @@
+// Функция из интернета по генерации случайного числа из диапазона
+// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+const getRandomInteger = (min = 0, max = 1) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+const isTaskExpired = (dueDate) => {
+  if (dueDate === null) {
+    return false;
+  }
+
+  let currentDate = new Date();
+  currentDate.setHours(23, 59, 59, 999);
+  currentDate = new Date(currentDate);
+
+  return currentDate.getTime() > dueDate.getTime();
+};
+
+const isTaskRepeating = (repeating) => {
+  return Object.values(repeating).some(Boolean);
+};
+
+const humanizeTaskDueDate = (dueDate) => {
+  return dueDate.toLocaleString(`en-US`, {day: `numeric`, month: `long`});
+};
+
+export {
+  getRandomInteger,
+  isTaskExpired,
+  isTaskRepeating,
+  humanizeTaskDueDate
+};
